+++ lib/plausible/stats/custom_props.ex
@@ -7,21 +7,6 @@ defmodule Plausible.Stats.CustomProps do
   use Plausible.ClickhouseRepo
   import Plausible.Stats.Base
 
-  @doc """
-  Returns a breakdown of event names with all existing custom
-  properties for each event name.
-  """
-  def props_for_all_event_names(site, query) do
-    from(e in base_event_query(site, query),
-      array_join: meta in fragment("meta"),
-      group_by: e.name,
-      select: {e.name, fragment("groupArray(?)", meta.key)},
-      distinct: true
-    )
-    |> ClickhouseRepo.all()
-    |> Enum.into(%{})
-  end
-
   def fetch_prop_names(site, query) do
     case Query.get_filter_by_prefix(query, "event:props:") do
       {"event:props:" <> key, _} ->